feat: add membershipFeeRepository with collectivity

+++ Ex/src/main/java/com/example/Ex/DTO/CreateMembershipFeeDTO.java
@@ -1,17 +1,19 @@
 package com.example.Ex.DTO;
 
+import java.math.BigDecimal;
+
 public class CreateMembershipFeeDTO {
     private String eligibleFrom;
     private String frequency;
-    private double amount;
+    private BigDecimal amount;
     private String label;
 
     public String getEligibleFrom() { return eligibleFrom; }
     public void setEligibleFrom(String eligibleFrom) { this.eligibleFrom = eligibleFrom; }
     public String getFrequency() { return frequency; }
     public void setFrequency(String frequency) { this.frequency = frequency; }
-    public double getAmount() { return amount; }
-    public void setAmount(double amount) { this.amount = amount; }
+    public BigDecimal getAmount() { return amount; }
+    public void setAmount(BigDecimal amount) { this.amount = amount; }
     public String getLabel() { return label; }
     public void setLabel(String label) { this.label = label; }
 }
